Move java code to other package
