+++ app/src/main/java/com/getsixtyfour/openvpnmgmt/android/core/VPNLaunchHelper.java
@@ -3,7 +3,7 @@
  * Distributed under the GNU GPL v2 with additional terms. For full terms see the file doc/LICENSE.txt
  */
 
-package de.blinkt.openvpn.core;
+package com.getsixtyfour.openvpnmgmt.android.core;
 
 import android.content.Context;
 import android.content.Intent;
